Fix invalid usage of pattern/includes on rule `orderby-clause`
Single-match rules should NOT have a patterns array. The patterns in this rule
have no effect:

  orderby-clause:
    match: \b(ORDER BY)\b\s*
    captures:
      '1': { name: keyword.operator.query.orderby.apex }
    patterns:
      - include: '#ordering-direction'
      - include: '#ordering-nulls'

To use the patterns array, we'd need a way to know when the pattern begins and ends.
Then we could use begin/end regexes (instead of a single match:) or define a capture
group right next to the ORDER BY string.

Reference:

* Additional docs about textmate grammars: http://www.apeth.com/nonblog/stories/textmatebundle.html
* Scripts to do some analysis on grammars: https://github.com/heyimalex/tmlanguage-analysis